python/requirements.txt
mysql
pandas
